[feat] creation of the default button and switch

+++ lib/modules/products/data/models/product_model.dart
@@ -24,7 +24,7 @@ class ProductModel {
     return ProductModel(
       id: json['id'],
       title: json['title'],
-      price: json['price'],
+      price: json['price'].toDouble(),
       description: json['description'],
       category: json['category'],
       image: json['image'],
